perf(web): drop AJV and the DB schema from the eager entry graph
The SPA's entry graph measured 87 assets / 1,330,250 bytes raw. Two heavy
dependencies were pulled in by fully static import chains even though they
only serve secondary screens.

AJV + semver (~144 kB raw): agent list -> package card -> run button ->
MissingConnectionsModal -> IntegrationConnectionPicker ->
@appstrate/core/integration -> @afps-spec/schema, which instantiates
Ajv2020 at module scope. @afps-spec/schema is an npm dependency, so the
chain is broken here instead: the 412-recovery modal is now React.lazy and
only mounts while it is open. The Suspense fallback reuses the same Modal
frame and title, so the dialog still opens instantly.

drizzle-orm + the 18 schema files (~76 kB raw): shared-types imported
runStatusEnum by value, and a pgEnum() cannot be elided by the bundler.
That shipped drizzle-orm and every table and column name -- llm_usage,
oauth_clients, end_users, api_keys, all of oidc.ts -- into a public asset.
The status literals now live in an import-free packages/db/src/run-status.ts
that the pgEnum, the Zod validator and shared-types all derive from, so
there is still exactly one list and the DB enum cannot drift from what the
client ships. module-chat's hand-maintained copy of the same list (kept
precisely to avoid drizzle) is deleted in favour of the real tuple.

favicon.svg is a 358 kB wrapper around two base64 PNGs and browsers
preferred it over the 2 kB PNG; the page now requests raster only.

Entry graph: 87 assets / 1,330,250 B raw / 393,278 B gzip
          -> 82 assets / 1,074,549 B raw / 324,330 B gzip  (-19.2% / -17.5%)
drizzle symbols and table names: gone from the entire bundle, not just the
eager graph.
