+++ apps/web/index.html
@@ -12,8 +12,11 @@
       document.documentElement.classList.add(t);
     })();
   </script>
+  <!-- Raster only, on purpose: public/favicon.svg is a 358 kB wrapper around two
+       base64 PNGs (light/dark variants), and every browser that supports SVG icons
+       preferred it over the 2 kB PNG — it was the single largest asset the page
+       requested. Do not re-add it without shrinking it to real vector art. -->
   <link rel="icon" type="image/png" href="/favicon-96x96.png" sizes="96x96" />
-  <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
   <link rel="shortcut icon" href="/favicon.ico" />
   <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
   <meta name="apple-mobile-web-app-title" content="Appstrate" />
